Change query cache policy (again).  Fixes issue 156.
This reverts to the original policy: if a query returns a result that
is also in the cache, the cached entity is substituted for the query
result.  This is consistent with key.get().  Also, if the cache
indicates the entity was deleted, the result is skipped.  OTOH, if the
result is not yet in the cache, it is put into the cache, to be
returned by later key.get() calls.  (All assuming caching is on for
the key involved, of course.)

BEWARE: On the HRD, if a non-ancestor query returns a stale result, it
is still cached.  If you don't want this, turn off caching,
e.g. qry.fetch(use_cache=False).